+++ components/ui/toast/useNimbusToast.ts
@@ -2,15 +2,23 @@ import { useCallback, useMemo } from "react";
 import Toast from "react-native-toast-message";
 
 type Variant = "success" | "error" | "info" | "warning";
+type Position = "top" | "bottom";
 
 export function useNimbusToast() {
   const show = useCallback(
-    (args: { variant: Variant; title: string; message?: string }) => {
+    (
+      args: {
+        variant: Variant;
+        title: string;
+        message?: string;
+        position?: Position;
+      }
+    ) => {
       Toast.show({
         type: args.variant,
         text1: args.title,
         text2: args.message,
-        position: "bottom",
+        position: args.position ?? "bottom",
       });
     },
     []
