+++ features/self-care/screens/BreathWorkDetailScreen.tsx
@@ -1,8 +1,10 @@
 import React, {
+  useCallback,
   useContext,
   useEffect,
   useLayoutEffect,
   useMemo,
+  useRef,
   useState,
 } from "react";
 import {
@@ -31,8 +33,12 @@ import {
   buildBreathWorkRouteParams,
   mapBreathworkDetail,
   hydrateBreathWorkDetail,
-  type BreathWorkRouteParams,
 } from "@/features/self-care/utils/breathworkLibrary";
+import {
+  createWellnessSessionLaunchKey,
+  launchWellnessSessionInBackground,
+} from "@/features/self-care/utils/wellnessSessionLaunch";
+import type { BreathWorkRouteParams } from "@/features/self-care/types/breathworkTypes";
 import type { ColorSet, Spacing, Typography } from "@/theme/types";
 
 type BreathWorkDetailParams = BreathWorkRouteParams;
@@ -88,6 +94,8 @@ export default function BreathWorkDetailScreen() {
   const [detail, setDetail] = useState(fallbackDetail);
   const [isLoading, setIsLoading] = useState(true);
   const [loadError, setLoadError] = useState<string | null>(null);
+  const [isStartingBreathWork, setIsStartingBreathWork] = useState(false);
+  const hasLaunchedBreathWorkRef = useRef(false);
 
   const styles = useMemo(
     () => styling(theme, spacing, typography),
@@ -146,16 +154,49 @@ export default function BreathWorkDetailScreen() {
     };
   }, [breathworkId, fallbackDetail]);
 
-  const handleStartBreathWork = () => {
+  useEffect(() => {
+    hasLaunchedBreathWorkRef.current = false;
+    setIsStartingBreathWork(false);
+  }, [breathworkId]);
+
+  const handleStartBreathWork = useCallback(() => {
+    if (isStartingBreathWork || hasLaunchedBreathWorkRef.current) {
+      return;
+    }
+
     if (Platform.OS !== "web") {
       void Haptics.selectionAsync().catch(() => {});
     }
 
+    hasLaunchedBreathWorkRef.current = true;
+    setIsStartingBreathWork(true);
+
+    const contentObjectId = Number(detail.id);
+    const launchKey = Number.isFinite(contentObjectId)
+      ? createWellnessSessionLaunchKey("breathwork")
+      : null;
+
+    if (Number.isFinite(contentObjectId) && launchKey) {
+      void launchWellnessSessionInBackground(launchKey, {
+        activity_type: "breathwork",
+        content_type: "wellness_content.wellnesscontent",
+        content_object_id: contentObjectId,
+        source: "manual",
+        metadata: {
+          entry_surface: "content_detail",
+          test_mode: true,
+        },
+      });
+    }
+
     router.push({
       pathname: ROUTES.AUTH.SELF_CARE_BREATHWORK_SESSION,
-      params: buildBreathWorkRouteParams(detail),
+      params: {
+        ...buildBreathWorkRouteParams(detail),
+        ...(launchKey ? { breathworkSessionLaunchKey: launchKey } : {}),
+      },
     });
-  };
+  }, [detail, isStartingBreathWork]);
 
   return (
     <ScreenView bgColor={theme.background} style={styles.screen}>
@@ -330,9 +371,11 @@ export default function BreathWorkDetailScreen() {
           </View>
 
           <NimbusButton
-            label="Play Breath Work"
+            label="Start Breath Work"
             onPress={handleStartBreathWork}
-            accessibilityLabel="Play Breath Work"
+            loading={isStartingBreathWork}
+            disabled={isStartingBreathWork}
+            accessibilityLabel="Start Breath Work"
             accessibilityHint="Starts the selected breathwork practice"
             leftIcon={
               <Ionicons
