+++ features/self-care/styles/breathwork/breathWorkSessionStyles.ts
@@ -1,0 +1,352 @@
+import { Platform, StyleSheet } from "react-native";
+
+import type { ColorSet, Spacing, Typography } from "@/theme/types";
+
+export const makeBreathWorkSessionStyles = (
+  theme: ColorSet,
+  spacing: Spacing,
+  typography: Typography
+) =>
+  StyleSheet.create({
+    screen: {
+      paddingHorizontal: spacing.md,
+      paddingTop:
+        Platform.OS === "ios"
+          ? spacing["xxl"] + spacing["xxl"] * 0.4
+          : spacing.xl,
+    },
+    root: {
+      flex: 1,
+    },
+    header: {
+      marginBottom: spacing.md,
+    },
+    scrollContent: {
+      paddingBottom: spacing.xl,
+      gap: spacing.lg,
+    },
+    heroCard: {
+      height: 300,
+      borderRadius: 32,
+      overflow: "hidden",
+      backgroundColor: theme.surface,
+      borderWidth: 1,
+      borderColor: theme.borderMuted ?? "rgba(255,255,255,0.06)",
+      shadowColor: theme.shadow,
+      shadowOpacity: 0.24,
+      shadowRadius: 18,
+      shadowOffset: { width: 0, height: 10 },
+      elevation: 8,
+    },
+    heroImage: {
+      width: "100%",
+      height: "100%",
+    },
+    heroGlowPrimary: {
+      position: "absolute",
+      top: -36,
+      right: -12,
+      width: 160,
+      height: 160,
+      borderRadius: 999,
+      opacity: 0.8,
+    },
+    heroGlowSecondary: {
+      position: "absolute",
+      bottom: -50,
+      left: -24,
+      width: 180,
+      height: 180,
+      borderRadius: 999,
+      backgroundColor: "rgba(255,255,255,0.1)",
+      opacity: 0.56,
+    },
+    heroCopy: {
+      position: "absolute",
+      left: 20,
+      right: 20,
+      bottom: 20,
+      gap: 6,
+    },
+    heroKicker: {
+      ...typography.smallCaption,
+      letterSpacing: 2.2,
+      color: "#D5DBC8",
+      textTransform: "uppercase",
+      opacity: 0.92,
+    },
+    heroTitle: {
+      ...typography.h1,
+      color: "#F4F2E8",
+    },
+    heroSubtext: {
+      ...typography.body,
+      color: "#E6E8D7",
+      maxWidth: 320,
+      opacity: 0.88,
+    },
+    loadingChip: {
+      flexDirection: "row",
+      alignItems: "center",
+      gap: 8,
+      paddingHorizontal: 10,
+      paddingVertical: 8,
+      borderRadius: 999,
+      marginTop: 12,
+      alignSelf: "flex-start",
+      backgroundColor: "rgba(7, 9, 7, 0.52)",
+      borderWidth: 1,
+      borderColor: "rgba(255,255,255,0.08)",
+    },
+    loadingChipText: {
+      ...typography.smallCaption,
+      color: theme.textPrimary,
+      letterSpacing: 0.8,
+    },
+    motionCard: {
+      borderRadius: 32,
+      borderWidth: 1,
+      borderColor: theme.border ?? theme.borderMuted,
+      alignItems: "center",
+      overflow: "hidden",
+      backgroundColor: theme.surface,
+      shadowColor: theme.shadow,
+      shadowOpacity: 0.18,
+      shadowRadius: 16,
+      shadowOffset: { width: 0, height: 8 },
+      elevation: 6,
+    },
+    phaseCard: {
+      minHeight: 284,
+      borderRadius: 28,
+      paddingTop: spacing.xl,
+      paddingHorizontal: spacing.xl,
+      paddingBottom: spacing.xl,
+      backgroundColor: theme.surface,
+      borderWidth: 1,
+      borderColor: theme.border ?? theme.borderMuted,
+      gap: spacing.md,
+      justifyContent: "space-between",
+    },
+    phaseCardPressed: {
+      transform: [{ scale: 0.99 }],
+    },
+    phaseCardDisabled: {
+      opacity: 0.84,
+    },
+    phaseTopRow: {
+      flexDirection: "row",
+      alignItems: "flex-start",
+      justifyContent: "space-between",
+      gap: spacing.md,
+      marginBottom: spacing.xs,
+    },
+    phaseTitleBlock: {
+      flex: 1,
+      gap: 8,
+    },
+    phaseLabel: {
+      ...typography.h2,
+      color: theme.textPrimary,
+      letterSpacing: -0.2,
+    },
+    phaseStartPrompt: {
+      ...typography.caption,
+      color: theme.textSecondary,
+      lineHeight: 18,
+    },
+    phaseCuePanel: {
+      borderRadius: 24,
+      paddingHorizontal: spacing.lg,
+      paddingVertical: spacing.lg,
+      backgroundColor: theme.surfaceMuted,
+      borderWidth: 1,
+      borderColor: theme.borderMuted ?? "rgba(255,255,255,0.08)",
+      gap: spacing.xs,
+      minHeight: 100,
+    },
+    phaseCueLabel: {
+      ...typography.smallCaption,
+      color: theme.textSecondary,
+      letterSpacing: 2,
+      textTransform: "uppercase",
+    },
+    phaseActions: {
+      alignItems: "center",
+      gap: 10,
+      minWidth: 160,
+      alignSelf: "center",
+      marginTop: spacing.md,
+      paddingTop: 2,
+    },
+    phaseStartBadge: {
+      flexDirection: "row",
+      alignItems: "center",
+      justifyContent: "center",
+      gap: 6,
+      borderRadius: 999,
+      paddingHorizontal: 16,
+      paddingVertical: 8,
+      borderWidth: 1,
+    },
+    phaseStartText: {
+      ...typography.smallCaption,
+      textTransform: "uppercase",
+      letterSpacing: 1.4,
+    },
+    phasePill: {
+      flexDirection: "row",
+      alignItems: "center",
+      justifyContent: "center",
+      borderWidth: 1,
+      borderRadius: 999,
+      paddingHorizontal: 12,
+      paddingVertical: 6,
+    },
+    phasePillText: {
+      ...typography.smallCaption,
+      textTransform: "uppercase",
+      letterSpacing: 1.4,
+    },
+    phaseCue: {
+      ...typography.body,
+      color: theme.textPrimary,
+      lineHeight: 24,
+      minHeight: 48,
+    },
+    metricRow: {
+      flexDirection: "row",
+      alignItems: "center",
+      justifyContent: "space-between",
+      gap: spacing.md,
+      marginTop: "auto",
+      paddingTop: spacing.sm,
+      paddingBottom: spacing.xs,
+    },
+    metricBlock: {
+      flex: 1,
+      alignItems: "center",
+    },
+    metricValue: {
+      ...typography.h3,
+      color: theme.textPrimary,
+    },
+    metricLabel: {
+      ...typography.caption,
+      color: theme.textSecondary,
+      marginTop: 4,
+    },
+    metricDivider: {
+      width: 1,
+      height: 38,
+      backgroundColor: theme.borderMuted ?? "rgba(255,255,255,0.08)",
+    },
+    sequenceCard: {
+      borderRadius: 28,
+      padding: spacing.lg,
+      backgroundColor: theme.surface,
+      borderWidth: 1,
+      borderColor: theme.border ?? theme.borderMuted,
+      gap: spacing.md,
+    },
+    sequenceLabel: {
+      ...typography.smallCaption,
+      color: theme.textSecondary,
+      letterSpacing: 2,
+      textTransform: "uppercase",
+    },
+    sequenceRow: {
+      flexDirection: "row",
+      flexWrap: "wrap",
+      justifyContent: "space-between",
+      rowGap: spacing.sm,
+    },
+    sequenceChip: {
+      width: "48%",
+      height: 92,
+      borderRadius: 20,
+      borderWidth: 1,
+      borderColor: theme.borderMuted ?? "rgba(255,255,255,0.08)",
+      backgroundColor: theme.surfaceMuted,
+      paddingHorizontal: 14,
+      paddingVertical: 12,
+      gap: 4,
+      justifyContent: "space-between",
+      flexShrink: 0,
+    },
+    sequenceChipSelected: {
+      borderWidth: 1,
+    },
+    sequenceChipLabel: {
+      ...typography.smallCaption,
+      textTransform: "uppercase",
+      letterSpacing: 1.4,
+    },
+    sequenceChipValue: {
+      ...typography.button,
+      fontSize: 14,
+      lineHeight: 18,
+      marginTop: 2,
+    },
+    errorCard: {
+      borderRadius: 24,
+      backgroundColor: "rgba(247, 196, 139, 0.08)",
+      borderWidth: 1,
+      borderColor: "rgba(247, 196, 139, 0.22)",
+      padding: spacing.md,
+      gap: spacing.xs,
+      marginTop: spacing.xs,
+      marginBottom: spacing.sm,
+    },
+    errorHeaderRow: {
+      flexDirection: "row",
+      alignItems: "center",
+      justifyContent: "space-between",
+      gap: spacing.sm,
+    },
+    errorTitle: {
+      ...typography.button,
+      color: "#F7C48B",
+    },
+    errorText: {
+      ...typography.body,
+      color: theme.textSecondary,
+      lineHeight: 22,
+    },
+    insightCard: {
+      borderRadius: 28,
+      padding: spacing.lg,
+      backgroundColor: theme.surface,
+      borderWidth: 1,
+      borderColor: theme.border ?? theme.borderMuted,
+      gap: spacing.sm,
+    },
+    insightLabel: {
+      ...typography.smallCaption,
+      color: theme.textSecondary,
+      letterSpacing: 2,
+      textTransform: "uppercase",
+    },
+    insightText: {
+      ...typography.body,
+      color: theme.textPrimary,
+      lineHeight: 24,
+    },
+    insightSubtext: {
+      ...typography.caption,
+      color: theme.textSecondary,
+      lineHeight: 18,
+    },
+    tipRow: {
+      flexDirection: "row",
+      alignItems: "flex-start",
+      gap: spacing.sm,
+      marginTop: spacing.xs,
+    },
+    tipText: {
+      flex: 1,
+      ...typography.caption,
+      color: theme.textSecondary,
+      lineHeight: 18,
+    },
+  });
